Move interfaces to folder
