+++ Vivel/Controllers/BaseCRUDController.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Vivel.Interfaces;
 using Vivel.Services;
 
 namespace Vivel.Controllers
