+++ Vivel/Interfaces/IDonationService.cs
@@ -6,7 +6,7 @@ using Vivel.Database;
 using Vivel.Model.Dto;
 using Vivel.Model.Requests.Donation;
 
-namespace Vivel.Services
+namespace Vivel.Interfaces
 {
     public interface IDonationService : IBaseCRUDService<DonationDTO, DonationSearchRequest, DonationInsertRequest, DonationUpdateRequest>
     {
