+++ Vivel/Interfaces/IHospitalService.cs
@@ -6,7 +6,7 @@ using Vivel.Database;
 using Vivel.Model.Dto;
 using Vivel.Model.Requests.Hospital;
 
-namespace Vivel.Services
+namespace Vivel.Interfaces
 {
     public interface IHospitalService : IBaseCRUDService<HospitalDTO, HospitalSearchRequest, HospitalUpsertRequest, HospitalUpsertRequest>
     {
