+++ Vivel/Startup.cs
@@ -12,6 +12,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Vivel.Database;
+using Vivel.Interfaces;
 using Vivel.Services;
 
 namespace Vivel
